move stuff into mgmt
